[user][BE] Add user authorization and token processing

+++ backend/src/controllers/jarController.ts
@@ -1,15 +1,15 @@
 import mongoose from 'mongoose';
-import type { Request, Response } from 'express';
+import type { Response } from 'express';
 import Jar from '../models/JarSchema.js';
 import type { IJar } from '../models/JarSchema.js';
 import { validationResult } from 'express-validator';
 import errorFormatter from '../utils/errorFormatter.js';
 import jarMapper from '../utils/jarMapper.js';
+import type { IUserRequest } from '../middleware/getUserFromToken.js';
 
-export const getJar = (req: Request, res: Response): void => {
+export const getJar = (req: IUserRequest, res: Response): void => {
   (async () => {
-    // TODO: Remove params user from request, instead use user's token
-    const userId: string = req.params.id;
+    const userId = req.user?._id;
 
     await Jar.find({ users: userId })
       .then(async (existingJars) => {
@@ -24,19 +24,19 @@ export const getJar = (req: Request, res: Response): void => {
   })();
 };
 
-export const createJar = (req: Request, res: Response): void => {
+export const createJar = (req: IUserRequest, res: Response): void => {
   const errors = validationResult(req).formatWith(errorFormatter);
 
   if (!errors.isEmpty()) {
     res.status(400).json({ error: errors.array() });
     return;
   }
-  // TODO: Remove params user from request, instead use user's token
+
   const jar: IJar = new Jar({
     name: req.body.name,
     color: req.body.color,
-    users: req.body.user,
-    owner: req.body.user
+    users: req.user?._id,
+    owner: req.user?._id
   });
 
   jar.save().then((newJar) => {
@@ -52,10 +52,12 @@ export const createJar = (req: Request, res: Response): void => {
  * @param req { name: 'JarName', color: 'red'}
  * @param res Update Object Jar
 */
-export const updateJar = (req: Request, res: Response): void => {
+export const updateJar = (req: IUserRequest, res: Response): void => {
   const jarId: string = req.params.id;
   const jarName: string = req.body.name;
   const jarColor: string = req.body.color;
+  const userId = req.user?._id;
+
   const errors = validationResult(req).formatWith(errorFormatter);
 
   if (!mongoose.Types.ObjectId.isValid(jarId)) {
@@ -70,7 +72,7 @@ export const updateJar = (req: Request, res: Response): void => {
 
   (async () => {
     await Jar.findOneAndUpdate(
-      { _id: jarId },
+      { _id: jarId, owner: userId },
       { $set: { name: jarName, color: jarColor } },
       { returnDocument: 'after' })
       .then((updatedJar) => {
@@ -86,18 +88,20 @@ export const updateJar = (req: Request, res: Response): void => {
   })();
 };
 
-export const deleteJar = (req: Request, res: Response): void => {
-  // TODO: Check jar owner before delete
-  const jarId: string = req.params.id;
+export const deleteJar = (req: IUserRequest, res: Response): void => {
+  (async () => {
+    const jarId: string = req.params.id;
+    const userId = req.user?._id;
 
-  Jar.findOneAndDelete({ _id: jarId })
-    .then((jar) => {
-      if (!jar) {
-        return res.status(404).json({ error: [{ msg: 'No jar found' }] });
-      }
-      return res.status(200).json({ result: 'You have successfully removed the jar' });
-    })
-    .catch((err) => {
-      return res.status(500).json({ error: [{ msg: err }] });
-    });
+    await Jar.findOneAndDelete({ _id: jarId, owner: userId })
+      .then(async (jar) => {
+        if (!jar) {
+          return res.status(404).json({ error: [{ msg: 'No jar found' }] });
+        }
+        return res.status(200).json({ result: 'You have successfully removed the jar' });
+      })
+      .catch((err) => {
+        return res.status(500).json({ error: [{ msg: err }] });
+      });
+  })();
 };
